requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
jinja2
python-multipart
psycopg2-binary
python-jose
passlib
